// src/app/components/interfaces/heroes.interface.ts
// Generated by https://quicktype.io

export interface Heroes {
    id?:               string;
    superhero:        string;
    publisher?:        Publisher;
    alter_ego:        string;
    first_appearance: string;
    characters:       string;
    description:      string;
    alt_img?:        string;
}

export enum Publisher {
    DCComics = "DC Comics",
    MarvelComics = "Marvel Comics",
}